demos/03_discord_integration/requirements.txt
discord.py>=2.0.0
llama-cpp-python>=0.2.0
